+++ src/components/organisms/PlayerBar/index.ts
@@ -1,0 +1,11 @@
+import { connect } from 'react-redux'
+
+import PlayerBar from './PlayerBar'
+
+const mapStateToProps = (state) => {
+  return {
+    songDetails:  state.songsReducer.songDetails
+  }
+}
+
+export default connect(mapStateToProps)(PlayerBar)