+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ## [2.1.2] - 2020.xx.xx
-## Added
+## Changed
+- Refactor PlayerBar and used molecules (PlayInfo, PlayerControls, VolumeControl)
+- Extract redux state away from molecules up into PlayerBar
+- Add test coverage to molecules
+- Refactor Artwork component
+
 
 
 ## [2.1.1] - 2020.10.09